fix(web): extract action-state from "use server" files for Next 16
Production runtime logs (prod deployment dpl_6xGDRi3mxvz4dZZd9T7BkNK6iLPC,
commit 9e36f1a) showed every POST to /grows/new and /settings returning 500
with `Error: A "use server" file can only export async functions`. Next 16
enforces that "use server" files may only export async functions; const
exports of `*ActionInitialState` constants triggered the runtime throw on
every server-action invocation. This was the real root cause of the audit's
"Server Components render" error — masking my earlier diagnosis.

- Extract every `*ActionInitialState` const + its result type to a sibling
  `action-state.ts` file. Affected: grows/actions.ts (1),
  grows/[id]/actions.ts (4), settings/actions.ts (3), plants/actions.ts (1).
- Re-export the result types from the original `actions.ts` files via
  `export type { ... }` — SWC erases type-only re-exports, so the runtime
  module still only exports async functions, and existing consumers that
  imported the type from `actions.ts` keep working.
- Update each form (grow-form, plant-form, archive-button, delete-button,
  stage-stepper, edit-form, settings-{profile,timezone,email}-form) to
  import the InitialState constant + type from the new `action-state.ts`.
- Add a check-imports.mjs guardrail that scans every "use server" file
  for non-async exports (const/let/var/enum/class/non-async function) and
  fails the validate gate if any appear. This regression can't recur
  silently — drift is caught before it reaches a deploy.

Validation: pnpm run validate (5/5), pnpm turbo run type-check lint test
(705/705 web tests), pnpm run security:audit all green.
