+++ apps/web/src/app/(app)/grows/[id]/edit/edit-form.tsx
@@ -5,11 +5,11 @@ import Link from "next/link";
 import { Button, buttonStyles } from "@/components/ui/button";
 import { FormErrorSummary } from "@/components/form-error-summary";
 import { useActionWithRecovery } from "@/lib/client/action-runner";
+import { updateGrowAction } from "../actions";
 import {
-  updateGrowAction,
   updateGrowActionInitialState,
   type UpdateGrowActionResult,
-} from "../actions";
+} from "../action-state";
 
 const FIELD_META = {
   name: { label: "Grow name", targetId: "grow-name" },
