+++ apps/web/src/app/(app)/plants/new/plant-form.tsx
@@ -6,11 +6,11 @@ import { Button, buttonStyles } from "@/components/ui/button";
 import { FormErrorSummary } from "@/components/form-error-summary";
 import { useActionWithRecovery } from "@/lib/client/action-runner";
 type GrowRecord = { id: string; name: string; stage: string | null };
+import { createPlantAction } from "../actions";
 import {
-  createPlantAction,
   createPlantActionInitialState,
   type CreatePlantActionResult,
-} from "../actions";
+} from "../action-state";
 import { MAX_BULK_PLANT_COUNT } from "../constants";
 
 const FIELD_META = {
