+++ apps/web/src/app/(app)/settings/settings-timezone-form.tsx
@@ -3,11 +3,11 @@
 import { useEffect, useRef, useState, useTransition } from "react";
 import { useRouter } from "next/navigation";
 import { Button } from "@/components/ui/button";
+import { updateTimezoneAction } from "./actions";
 import {
-  updateTimezoneAction,
   updateTimezoneActionInitialState,
   type UpdateTimezoneActionResult,
-} from "./actions";
+} from "./action-state";
 
 const inputClassName =
   "mt-2 block w-full rounded-[1.15rem] border border-border/80 bg-surface px-4 py-3 text-sm text-foreground shadow-soft transition focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-accent/35";
